Fix log rotation

+++ pi-setup/setup.sh
@@ -3,14 +3,25 @@
 HOST="pi@192.168.0.139"
 PORT="1111"
 
-ssh -t $HOST "sudo cp /etc/ssh/sshd_config /tmp"
-ssh -t $HOST "sudo sed -i 's/#Port 22/#Port 22\nPort 1111/' /etc/ssh/sshd_config"
-ssh -t $HOST "sudo service ssh restart"
+# Add SSH Keys
+if [[ ! -f authorized_keys ]]; then
+    echo You need to create the file authorized_keys and insert public keys into it.
+    exit 1
+fi
+
+ssh -t $HOST -p $PORT "mkdir -p ~/.ssh && chmod 700 ~/.ssh"
+scp -P $PORT authorized_keys $HOST:/home/pi/.ssh/authorized_keys
+ssh -t $HOST -p $PORT "chmod 600 /home/pi/.ssh/authorized_keys"
+
+# Secure PI settings
+ssh -t $HOST -p $PORT "sudo cp /etc/ssh/sshd_config /tmp"
+ssh -t $HOST -p $PORT "sudo sed -i 's/#Port 22/#Port 22\nPort 1111/' /etc/ssh/sshd_config"
+ssh -t $HOST -p $PORT "sudo service ssh restart"
 
 ssh -t $HOST -p $PORT "sudo apt install -y unattended-upgrades"
 ssh -t $HOST -p $PORT "mkdir -p /tmp/secure-pi"
 
-scp 02periodic $HOST:/tmp/secure-pi
+scp -P $PORT 02periodic $HOST:/tmp/secure-pi
 ssh -t $HOST -p $PORT "sudo chown root:root /tmp/secure-pi/02periodic"
 ssh -t $HOST -p $PORT "sudo mv /tmp/secure-pi/02periodic /etc/apt/apt.conf.d"
 
@@ -22,3 +33,7 @@ ssh -t $HOST -p $PORT "sudo ufw allow from 192.168.0.0/24 to any port $PORT"
 ssh -t $HOST -p $PORT "sudo ufw allow from 192.168.1.0/24 to any port $PORT"
 ssh -t $HOST -p $PORT "sudo ufw enable"
 ssh -t $HOST -p $PORT "sudo ufw status verbose"
+
+# Other setup
+scp -P ${PORT} setup-path.sh $HOST:/tmp/setup-path.sh
+ssh -t $HOST -p $PORT "/tmp/setup-path.sh"
